+++ ppci/arch/mos6500/target.py
@@ -4,8 +4,10 @@ from .instructions import isa
 
 
 class Mos6500Target(Target):
-    def __init__(self):
-        super().__init__('6500')
+    name = '6500'
+
+    def __init__(self, options=None):
+        super().__init__(options=options)
         self.isa = isa
         self.assembler = BaseAssembler()
         self.assembler.gen_asm_parser(isa)